requirements.txt
streamlit
pillow
requests
ollama